Simplified lazy_string, renamed some methods, moved common code to transformer/common, reduced unnecessary copies
